- added sql - dummy data - removed cloudtype entity and winddirection entity
- cloudtype and winddirection classes are removed and are strings
instead , and are in weatherinfo
- changed weatherInforepository interface from integer to Long
- added empty constructors for weatherstation and weatherinfo , spring
is complaining otherwise and produces errors
- weatherstation and weatherino , when generated creates an assosiative
table
- added sql insert for dummy data

+++ src/main/java/repository/WeatherInfoRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 /**
  * Created by Liz on 2017-02-09.
  */
-public interface WeatherInfoRepository extends CrudRepository <WeatherInfo, Integer> {
+public interface WeatherInfoRepository extends CrudRepository <WeatherInfo, Long> {
 
 }
